Created new component for UI and added it in main component also removed save button
