--- src/App.css
@@ -1,76 +1,0 @@
-main {
-	height: 100%;
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	position: absolute;
-	background-color: #beff00;
-}
-
-h1 {
-	font-size: 40px;
-	font-weight: 600;
-	color: #263200;
-	flex: 2;
-	padding-top: 20px;
-}
-
-.counter-text {
-	font-size: 30px;
-	text-align: center;
-	text-decoration: none;
-	border: none;
-	padding: 10px;
-	background-color: #beff00;
-	color: #263200;
-	border-radius: 5px;
-	outline: none;
-	border-bottom: solid;
-}
-
-.counter-text::placeholder {
-	font-size: 25px;
-}
-.counter-text:hover {
-	border-bottom: 5px solid;
-}
-
-.counter-count {
-	font-size: 80px;
-	padding: 50px;
-	color: #263200;
-	position: relative;
-	top: 120px;
-	flex: 2;
-}
-
-.button-container {
-	display: flex;
-	height: 300px;
-	width: 374px;
-	flex: 1;
-}
-
-button {
-	font-size: 50px;
-	border: solid white 1px;
-	cursor: pointer;
-	background-color: #263200;
-	color: #beff00;
-	flex: 1;
-	justify-content: center;
-	align-items: center;
-	margin-bottom: 20px;
-}
-
-button:hover {
-	opacity: 0.5;
-	color: white;
-}
-
-.button-subtract:disabled {
-	background-color: darkolivegreen;
-	opacity: 1;
-}
