+++ src/index.css
@@ -1,6 +1,5 @@
-body{
-    margin: 0px;
-    padding: 0px;
-    overflow: hidden;
+body {
+	margin: 0px;
+	padding: 0px;
+	overflow: hidden;
 }
-
